fix(stagewise): fix notification-placement
